Don't use size_t in f3kdb.h, since that needs another header file

+++ include/f3kdb.h
@@ -93,6 +93,6 @@ F3KDB_API(int) f3kdb_params_fill_by_string(f3kdb_params_t* params, const char* p
 F3KDB_API(int) f3kdb_params_fill_preset(f3kdb_params_t* params, const char* preset, int interface_version = F3KDB_INTERFACE_VERSION);
 F3KDB_API(int) f3kdb_params_sanitize(f3kdb_params_t* params, int interface_version = F3KDB_INTERFACE_VERSION);
 F3KDB_API(int) f3kdb_video_info_sanitize(f3kdb_video_info_t* vi, int interface_version = F3KDB_INTERFACE_VERSION);
-F3KDB_API(int) f3kdb_create(const f3kdb_video_info_t* video_info, const f3kdb_params_t* params, f3kdb_core_t** core_out, char* extra_error_msg = nullptr, size_t error_msg_size = 0, int interface_version = F3KDB_INTERFACE_VERSION);
+F3KDB_API(int) f3kdb_create(const f3kdb_video_info_t* video_info, const f3kdb_params_t* params, f3kdb_core_t** core_out, char* extra_error_msg = nullptr, int error_msg_size = 0, int interface_version = F3KDB_INTERFACE_VERSION);
 F3KDB_API(int) f3kdb_destroy(f3kdb_core_t* core);
 F3KDB_API(int) f3kdb_process_plane(f3kdb_core_t* core, int frame_index, int plane, unsigned char* dst_frame_ptr, int dst_pitch, const unsigned char* src_frame_ptr, int src_pitch);
